+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [3.0.2] - 2025-01-25
+
+### Fixed
+- Fixed `RouteDefinition` assertion error in `_runGuards()` when evaluating guards
+  - Changed `RouteDefinition(path: location)` to `RouteDefinition.stub(location)`
+  - This was causing crashes when authenticated users navigated to protected routes
+
+---
+
 ## [3.0.1] - 2025-01-25
 
 ### Fixed
@@ -117,6 +126,7 @@ See the [README](README.md) for the progressive migration guide.
 
 ---
 
+[3.0.2]: https://github.com/chekarhamza88-stack/nav_bridge/releases/tag/v3.0.2
 [3.0.1]: https://github.com/chekarhamza88-stack/nav_bridge/releases/tag/v3.0.1
 [3.0.0]: https://github.com/chekarhamza88-stack/nav_bridge/releases/tag/v3.0.0
 [2.0.0]: https://github.com/chekarhamza88-stack/nav_bridge/releases/tag/v2.0.0